termodinamica/requirements.txt
pyserial==3.5
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
